Hoist test files
