+++ src/components/feedback/Notice.test.tsx
@@ -3,7 +3,7 @@ import { vi, describe, beforeEach, it, expect } from "vitest"
 import { screen, render } from "@testing-library/react"
 import userEvent from "@testing-library/user-event"
 
-import { Notice } from "../Notice"
+import { Notice } from "./Notice"
 
 describe("Notice", () => {
   const onClose = vi.fn()
