+++ src/components/forms/PhoneInput.test.tsx
@@ -3,7 +3,7 @@ import { vi, beforeEach, describe, it, expect } from "vitest"
 import { screen, render } from "@testing-library/react"
 import userEvent from "@testing-library/user-event"
 
-import { PhoneInput, CountryCode } from "../PhoneInput"
+import { PhoneInput, CountryCode } from "./PhoneInput"
 
 describe("PhoneInput", () => {
   const onChange = vi.fn()
