+++ src/components/overlay/FullscreenLoading.test.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import { screen, render } from "@testing-library/react"
 
-import { FullscreenLoading } from "../FullscreenLoading"
+import { FullscreenLoading } from "./FullscreenLoading"
 
 describe("FullscreenLoading", () => {
   beforeEach(() => {
